% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{morph_fun}
\alias{morph_fun}
\title{Generic morphology function}
\usage{
morph_fun(species, length_m, morph)
}
\arguments{
\item{species}{a vector of species codes}

\item{length_m}{a vector of lengths in meters}

\item{morph}{name of the morphological measurement (length one character vector)}
}
\value{
vector of measurements
}
\description{
Generic morphology function
}
