% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{power_law}
\alias{power_law}
\title{Power law}
\usage{
power_law(a, b, x)
}
\arguments{
\item{a}{intercept of the log10-log10 relationship}

\item{b}{slope of the log10-log10 relationship}

\item{x}{untransformed values for power law calculation}
}
\value{
a vector of power law results
}
\description{
Power law
}
