% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{rorq_bizygomatic}
\alias{rorq_bizygomatic}
\title{Rorqual bizygomatic skull width}
\usage{
rorq_bizygomatic(species, length_m)
}
\arguments{
\item{species}{a vector of species codes}

\item{length_m}{a vector of lengths in meters}
}
\value{
a vector of bizygomatic skull widths in m
}
\description{
Rorqual bizygomatic skull width
}
\examples{
# A 22m blue whale
rorq_bizygomatic("bw", 22)

# A 7m minke
rorq_bizygomatic("ba", 7)
}
