% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{rorq_engulf}
\alias{rorq_engulf}
\title{Rorqual engulfment capacity}
\usage{
rorq_engulf(species, length_m)
}
\arguments{
\item{species}{a vector of species codes}

\item{length_m}{a vector of lengths in meters}
}
\value{
a vector of engulfment capacities in kg of water
}
\description{
Rorqual engulfment capacity
}
\examples{
# A 22m blue whale
rorq_engulf("bw", 22)

# A 7m minke
rorq_engulf("ba", 7)
}
