% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{rorq_flipper}
\alias{rorq_flipper}
\title{Rorqual flipper length}
\usage{
rorq_flipper(species, length_m)
}
\arguments{
\item{species}{a vector of species codes}

\item{length_m}{a vector of lengths in meters}
}
\value{
a vector of flipper lengths in m
}
\description{
Rorqual flipper length
}
\examples{
# A 22m blue whale
rorq_flipper("bw", 22)

# A 7m minke
rorq_flipper("ba", 7)
}
