% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{rorq_fluke}
\alias{rorq_fluke}
\title{Rorqual fluke length}
\usage{
rorq_fluke(species, length_m)
}
\arguments{
\item{species}{a vector of species codes}

\item{length_m}{a vector of lengths in meters}
}
\value{
a vector of fluke lengths in m
}
\description{
Rorqual fluke length
}
\examples{
# A 22m blue whale
rorq_fluke("bw", 22)

# A 7m minke
rorq_fluke("ba", 7)
}
