% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{rorq_mandible}
\alias{rorq_mandible}
\title{Rorqual projected mandible length}
\usage{
rorq_mandible(species, length_m)
}
\arguments{
\item{species}{a vector of species codes}

\item{length_m}{a vector of lengths in meters}
}
\value{
a vector of laterally projected mandible lengths in m
}
\description{
Rorqual projected mandible length
}
\examples{
# A 22m blue whale
rorq_mandible("bw", 22)

# A 7m minke
rorq_mandible("ba", 7)
}
