% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{rorq_mass}
\alias{rorq_mass}
\title{Rorqual mass}
\usage{
rorq_mass(species, length_m)
}
\arguments{
\item{species}{a vector of species codes}

\item{length_m}{a vector of lengths in meters}
}
\value{
a vector of masses in kg
}
\description{
Rorqual mass
}
\examples{
# A 22m blue whale
rorq_mass("bw", 22)

# A 7m minke
rorq_mass("ba", 7)
}
