% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{rorq_massratio}
\alias{rorq_massratio}
\title{Rorqual engulfed water mass to body mass ratio}
\usage{
rorq_massratio(species, length_m)
}
\arguments{
\item{species}{a vector of species codes}

\item{length_m}{a vector of lengths in meters}
}
\value{
a vector of ratios (engulfed water mass to body mass)
}
\description{
Rorqual engulfed water mass to body mass ratio
}
\examples{
# A 22m blue whale
rorq_massratio("bw", 22)

# A 7m minke
rorq_massratio("ba", 7)
}
