% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{rorq_vgb}
\alias{rorq_vgb}
\title{Rorqual ventral groove blubber length}
\usage{
rorq_vgb(species, length_m)
}
\arguments{
\item{species}{a vector of species codes}

\item{length_m}{a vector of lengths in meters}
}
\value{
a vector of VGB lengths in m
}
\description{
Rorqual ventral groove blubber length
}
\examples{
# A 22m blue whale
rorq_vgb("bw", 22)

# A 7m minke
rorq_vgb("ba", 7)
}
